Move block components and directories
