+++ src/modules/block/components/blockDetails/blockDetails.css
