+++ src/modules/block/components/blockDetails/blockDetails.js
