+++ src/modules/block/components/blockDetails/blockDetails.test.js
