+++ src/modules/block/components/blocksOverview/index.js
@@ -1,0 +1,1 @@
+export { default } from './blocksOverview';
