+++ src/modules/block/manager/blockDetailsManager.js
@@ -6,7 +6,7 @@ import { withRouter } from 'react-router-dom';
 import { getBlock } from '@block/utils';
 import withData from '@common/utilities/withData';
 import { selectSearchParamValue } from 'src/utils/searchParams';
-import BlockDetails from './blockDetails';
+import BlockDetails from '../components/blockDetails';
 
 const mapStateToProps = (state, ownProps) => ({
   id: selectSearchParamValue(ownProps.history.location.search, 'id'),
